requirements.txt
alembic==1.12.1
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
coverage==7.8.2
dnspython==2.7.0
email-validator==2.1.0
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fpdf==1.7.2
greenlet==3.0.3
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.3.0
packaging==25.0
pandas==2.3.0
pluggy==1.6.0
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.23
text-unidecode==1.3
typing_extensions==4.14.0
tzdata==2025.2
Werkzeug==2.3.7
WTForms==3.2.1
pg8000==1.30.3
gunicorn==21.2.0
